# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Documentação - Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/auth/"
        }
    ],
    "paths": {
...
        "/vinculo/store": {
            "post": {
                "tags": [
                    "Vinculo"
                ],
                "summary": "Vinculo store",
                "operationId": "08ce6c8545164a148796a952c28aef94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pac_codigo",
                                    "plan_codigo",
                                    "nr_contrato"
                                ],
                                "properties": {
                                    "pac_codigo": {
                                        "type": "number"
                                    },
                                    "plan_codigo": {
                                        "type": "number"
                                    },
                                    "nr_contrato": {
                                        "type": "string",
                                        "example": "148956"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/vinculo/update/{id}": {
            "put": {
                "tags": [
                    "Vinculo"
                ],
                "summary": "Vinculo update",
                "operationId": "0bb8306b0147068f3a39217a53abb392",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pac_codigo": {
                                        "type": "number"
                                    },
                                    "plan_codigo": {
                                        "type": "number"
                                    },
                                    "nr_contrato": {
                                        "type": "string",
                                        "example": "148956"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/vinculo/delete/{id}": {
            "delete": {
                "tags": [
                    "Vinculo"
                ],
                "summary": "Vinculo delete",
                "operationId": "dbb163a4641eb4778dbf81e2835884c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/vinculo/show": {
            "get": {
                "tags": [
                    "Vinculo"
                ],
                "summary": "Vinculo show",
                "operationId": "8bd0895d9284f618a64b5df1458a0a59",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Para realizar login adione Bearer + o token. Exemplo: Bearer hjfdsfkhjhjerryue",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}